requirements-dev.txt
Django~=3.2.6
dj-database-url==0.5.0
dj-static==0.0.6
python-decouple==3.3
django-allauth
